Extract wizard controller into controllers/wizard.rs
Move the remaining wizard-related callback wiring into a single
controller module: on_step_changed, on_item_activated, on_toggle,
on_select_confirmed, on_text_confirmed, on_select_filter_changed,
on_text_input_edited (password strength), on_key_nav_*, the
handle_item_activated dispatcher, and the show_select / show_text_input
/ show_users_popup / show_string_list / show_package_search helpers.
Also fold the duplicated locale/keyboard fuzzy filter into a single
fuzzy_filter helper.

main.rs is now ~157 lines (down from ~2110): mod declarations, CLI
parsing, run_gui orchestration that calls each controller's setup(),
and the on_quit_requested handler. Each controller is self-contained
and only touches its own Slint global plus the shared GlobalConfig and
EditingModels.
