+++ slint-ui/src/controllers/mod.rs
@@ -1,3 +1,4 @@
 pub mod install;
 pub mod lists;
 pub mod welcome;
+pub mod wizard;
